Move base activity functionality into DropInActivity and remove BaseActivity superclass.
